requirements.txt
asgiref==3.7.2
better-profanity==0.7.0
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-jet-reboot==1.3.3
django-localflavor==4.0
django-multiselectfield==0.1.12
django-oauth-plus==2.2.9
django-oauth2-provider==0.2.6.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
httpie==3.2.2
httplib2==0.22.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oauth2==1.9.0.post1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-decouple==3.8
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==1.0.0
rich==13.5.2
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
whitenoise==6.5.0
wrapt==1.15.0
